Fixes Concord configuration validation bug occurring when using local loopback.
In particular, this commit fixes a bug where Concord's logic to validate that local loopback (i.e. 127.0.0.1) is used for hosts for which it is expected could trigger an exception and crash Concord during its boot-up by attempting to determine which node a configuration file belongs to before enough information has been loaded into the config to determine this.