[VB-1303] Create VSphereOrchestrator for on-premise provisioning
This patch creates a VSphereOrchestrator that implements the
Orchestrator contract, in order for provisioning service to provision to
vCenter environment.

The patch allows VSphereOrchestrator to create a VSphereHttpClient with
an insecure (i.e. trust self-signed certificate) SSL context, since
on-premise vCenter may not have a valid CA-signed certificate signing
over its IP or FQDN (which is usually on corporate network).