Create functions for tests to be able to collect and clean deployment events.
This change collects functionality from the Persephone tests and wraps them up into
simple calls that can be made from other places in the test framework.

It also adds one test to the HelenAPITests suite which uses it.  This test will be
skipped for the time being because of the time it takes and intermittent failures
to deploy or clean up.