Improve Concord config generation for docker env
This idea is, that most developers won't have a self-build
(concord-core:latest) image but rather pull it from artifactory. In either
case, docker-compose uses the `.env` file anyways to determine the proper
images. In this change, the config generation script will use the same file to
find the Concord image in order to run the configuration generation utility.