Place default logger configuration in Concord Docker image.
This commit ammends the Concord Dockerfile so that a default log4cplus configuration is included directly in the Docker image; it also changes the default value for the logger_config configuration parameter to point to this default file. It also includes a related change to logger configuration in our docker-compose dev environment so all 4 Concord nodes there will share 1 log configuration file for developer convenience.