Allow Persephone to use default value for logger_config.
We have pending changes to Concord out for review to add a default log4cplus configuration to Concord's Docker image. In order to have Concord use this default logger config, Persephone should not specify logger_config in the Concord configuration input so that the configuration generator can fill in the path to the default logger config. This commit makes that change. Explicit specifications of logger_config are also removed from sample configuration files contained in the persephone directory.