Populate user set page size properly
While playing around with the page size I hit an assert which verifies that the
assumption about the page size in different parts of the system are the same.
When populating the page size to the state transfer config I forgot to also
populate it to the replica config which later will turn into the
ReplicaConfigSingleton which is used to retrieve global config values.