Concord: Optional global config values
This change introduces a new tag that allows for optional global config values
for Concord. The motivation comes from recently introduces SBFT nobs. Their
default values are part of concord-bft and shouldn't be overwritten by Concord.
Hence, if the user doesn't set those values we use the default from concord-bft.
So far, the default values were matching the concord-bft values but we don't
want to maintain that compatibility.