[VB-980] Allow Concord VM to connect to insecure container registry
This patch modifies CloudInitConfiguration (executed on the Concord VM
during initial boot) to modify the VM's Docker engine settings to allow
non-HTTPS traffic connection to a specified container registry (passed
in via provisioning parameters). The modification to the Docker engine
startup parameter is only done if the provisioning parameter specified a
HTTP URL instead of HTTPS - the default is still TLS secure connection.

This is to allow setups like labs and testing where the premise may not
have access to CA-cert signed container registry hosts, which is
especially true for walled-garden type of deployments where there is no
connectivity to public internet.