[VB-1183] Create fleet-service packaging Dockerfile
This patch adds a containerized build Dockerfile for fleet-service
container image.

The patch also updated a long-outstanding package that could not be
updated along with the rest of Kotlin version (atomicfu) due to its
dependent build plugins not residing within VMware artifactory. The
ticket to mirror Kotlin's Bintray repository was finally fulfilled
recently, allowing atomicfu to be properly version-updated.