Remove explicit reference to gRPC's CallOptions
This patch removes the current usages of gRPC'S CallOptions to allow
code under various "src/common" to remain true to "platform-agnostic".
Otherwise it would be necessary to create "src/jvm" and expect/actual
declarations for the corresponding Kotlin code - in order to comply with
standard MPP conventions.

This patch is needed in order to bring in additional static lint tools
that catch multi-platform Kotlin (MPP) compliance.