Update Persephone Kotlin dependencies
This patch moves forward Kotlin dependencies to 1.4.50.

While updating the dependencies forward, it was discovered that the
reactive stream publisher support (from Kotlin coroutines) have
tightened to not allow a coroutine context with explicitly attached
Job to be passed in as argument. As a result of fixing the existing
code to adapt to the new restrictions, a latent bug was discovered
in IteratingPublisher.

The fix was to explicitly check for job cancellation in the event that
subscriber subscribes to an IteratingPublisher *after* the publisher
instance have already been close()'ed.