Upload DAR packages
This change is one step closer to a better test setup for DAML. We remove the
test tool from the Ledger API docker file and upload the test DAR packages at
runtime. The `upload_dar.py` file is an executable as well as a library which
is being used by the Hermes test.
The gRPC/Protobuf code is generated from the DAML v1 API as described in the
README. I only had to adjust the import but the rest of the generated code is
untouched.