Management service for DAML
With this change we're starting two services on Concord. One gRPC server for
DAML requests and one API server (pure protobuf messages) for management. The
latter supports peer requests only due to it also contains Ethereum API
endpoints which cannot be exposed.
In addition, this change provides a `conc_getpeers` executable which is used to
test the PeerRequest API on DAML and Eth deployments.