Remove usage of fixed32 in Persephone API proto definition
This patch makes a wire-incompatible change to Persephone API model
definition (but fully compatibly with existing Helen due to Helen's
consumption of existing API Persephone bindings).

There is a limitation in the generated source bindings for Kotlin source
files in that it's not possible to annotate a type parameter (e.g.
`Map<Int, Int>` where Int needs an additional annotation). As a result,
it's not possible to convert proto definition such as `map<int, int>` to
its JVM equivalent without losing the encoding intent for the `int`
fields to be fixed encoding instead of varint.

The effort to modify the code generator to emit valid JVM bindings for
`map<int, int>` is non-trivial and involves using additional runtime
types to encapsulate a fixed32 / fixed64 protobuf type (resulting in
additional encoding overhead).

Given that Helen is not impacted by this incompatible change, the only
consumer of the effected API is Hermes (testing). Existing test code has
not extracted / extended usage of the affected protobuf fields. Making a
one-time change to use fixed32 to int32 seems to be the most prudent
action to take currently.