Improve DAML tests
The test tool image now contains JAVA in order to extract the DAR files and
run the tool. With this image in place we can simplify the test and only
donwload the DAR files instead of the full tool. After uploading those files
we run the ledger api test tool from within the test tool container.