Ethereum config flag
This change uses the Ethereum config flag, similar to HLF and DAML which
exist already. We want to make sure that one and only of those features is
enabled at runtime and in order to not break exisiting workflows we set this
flag to `true by default`.
In addition, this change also spawns the HLF service in a separate thread in
order to also start the management service (which for now, is only the list of
Concord nodes).