DAML Ledger API v21: Integrate with CI/CD
This change updates the docker-compose-daml scenario to simulate the real world
deployment with one indexer per committer node (ledger api + concord + damle).
I ran into issues with docker and keyservers, hence the DAML dockerfiles now
pull sbt directly from bintray which is similar to what we do for the test
tool already. In addition, I removed the test tool from `build.sbt` as well
due to we're using/downloding it in a separate container.
Prior to this change, I updated the test tool in artifactory to version 21.