Enhance protobuf module to skip encoding default values
This patch enhances the forked Kotlin serialization protobuf encoder to
allow skipping of fields when the value to be encoded on the wire is the
same as field's scalar type value.

Specifically, the patch:
1. Skips encoding scalar numeric values if the value to be encoded is 0.
2. Skips encoding byte arrays and strings if the array length is 0.
3. Skips encoding message / struct if the overall payload is 0.

The end result brings the current protobuf encoder to be semantically
equivalent to the official Google Java implementation of protobuf in
terms of what is encoded on the wire for proto3 schema definitions.

The patch also fixes a bug in the code generator where @Required was not
annotated on required fields (when defined in proto2 format).