Add count gauge for each timing stat, render on graph
The count is redundant in many cases, because many of these stats are
always updated together, but the redundancy means you don't have to
remember which stats go with which count.

When we're actually adding non-timing stats, separate counts may make
sense, but for now I have removed them, and use just the hdrhistogram
count instead.