[VB-1556] Fix Persephone KV persistence unit-test failure
This patch follows up previously committed diagnostic patch to expose
the root-cause of KV publish event stream not terminating once KV server
close() is invoked.

The reason is because an onError() is signaled in the client subscriber
due to an internal send channel being already closed (since server was
explicitly shutdown). The test made an incorrect expectation that server
close() should only translate to onComplete() being signaled eventually
for all clients.

This patch fixes the test expectation with respect to termination
condition (the test still expects signal observations to be according
to the original test specification).

In addition, this patch improves BroadcastingPublisher (used internally
by KV server) to allow waitForSubscription() to take on an optional
waitCount parameter. Unit-tests (as well as future product use-cases)
can leverage this parameter to precisely determine the condition to
which the wait-task should return (i.e. some specified number of
subscribers to have subscribed).

Lastly, the patch adds two unit-tests to test BroadcastingPublisher
behavior.