[VB-1556] Fix Persephone KV persistence unit-test failure (part 2)
This patch follows up previously committed fix, and enhances the fix to
avoid the potential of "blocking" subscriber coroutines from exiting due
to not being able to send to BroadcastingPublisher's internal actor
that keeps track of subscriber count. This is due to the internal
command channel being left open (for the prior fix) while the actor
coroutine was being told to cancel/close.