Make Hermes start the DB (but not initialize it).
Since Hermes stopped stopped starting and initializing the DB, leaving
the task solely to docker-compose, we have seen intermittent failures
to connect to the DB to initialize it.  One difference is that Hermes
would wait until there was a DB connection, then initialize it, while
docker-compose starts sending SQL as soon as the container is up.

This change only puts back the Hermes code that managed bringing up
and stopping the DB container.  It does not re-introduce running
of the SQL to initialize the DB.  Let's see if that fixes it.