VB-1605: Do not copy & create/destroy Slivers during RocksDB Compare
Performance data showed that we were spending a lot of time in
concord::storage::RocksKeyComparator::Compare, and that much of that
time was spent either copying (copyRocksdbSlice) or constructing and
destructing Sliver objects. This PR adds implementations of the few
functions needed for this comparison, that take bare byte
pointers. This allows us to pass the internal pointer from either a
rocksdb::Slice or a concord::consensus::Sliver without having to
translate between the two.