Revert VB-1605 / MR !700, no-copy rocks comparator
This change broke config-service, because it added a dependency from
conc_genconfig to secp256k1, which is not on the path during config
service's execution.

The VB-1605 change has been ported to the concord-bft, where this
storage code has been moved. It will come back when we update
concord-bft versions. In the concord-bft repo, it did not add a
dependency on secp256k1.

This reverts commit 2dbe71cc2ccd37eeebb8b6d27252089797e70332, reversing
changes made to 5d19773b7b07e778298f4d06db30066eb8b1cef7.