Concord: Generate time_source_id
With time service enabled we also need to specify a name for each time source.
Our default is the system time for each Concord node. Given that these are
internal sources, this change proposes to keep the name internal as well and
hence generate it at config generation time. Meaning, the concord user doesn't
need to come up with names in order to enable time service. Practically, in our
dev environment, that mean that ConfigService (part of deployment) doesn't need
to be adapted beyond enabling time service.