[VB-1621] Handle orchestration site validation error gracefully
This patch fixes an issue where if the validation workflow has timed
out, the client caller does not receive any indication and ends up
hanging for a response.

The patch also introduces a unit-test (and a fault-injection condition)
to test the handling of general handling of errors raised to
OrchestrationSiteService.