VB-1605: update concord-bft to copy less in rocksdb iterator
Changes that were previously made in this repo, to make fewer
unnecessary copies in our rocksdb comparator, were moved to
concord-bft in this commit:
https://github.com/vmware/concord-bft/pull/223. This PR updates our
concord-bft dependency to pull in that work.

This also picks up other work to reduce Sliver
allocation/deallocation, and increase safety around modifications of
those buffers. This is the reason for the addition of `const` and `&`
in KVB interfaces. (https://github.com/vmware/concord-bft/pull/227)