Limit the amount of gRPC threads for DAML
The performance team hit an issue with concord-bft which put the system into a
live-lock. Hence, incoming requests couldn't be processed because there was no
concord-bft client available. This change limits the amount of threads the gRPC
server will use depending on the number client proxies per replica.

Testing:
Before, I flooded my test setup with API requests and hit ~800 threads but
concord-bft continued to work and hence it finished the work eventually.
After, the number of threads in the concord process doesn't reach 40.