Run perf team's benchmark
We had two parallel "ballot" performance tools. Remove the old one,
and make hermes run the one that the performance team's bencmark tool
runs.

This version still runs the performance*.jar directly, instead of
dapp-bench*.jar. The latter runs instances of the former anyway. It
seemed simpler to have hermes run the instance directly, instead of
writing a config file for the bench tool. This required slight changes
to the instance tool to support this, since hermes is running the jar
instead of loading the jar and calling methods.