Enabled BFT debug statistics during jenkins performance run
This will cause concord to print once every seven seconds a summary like this:

2019-09-25T17:36:14.719 [139970945730304] INFO  concord.bft %%  STAT:
    ReadOnlyThroughput = 0.00
    WriteThroughput = 5.14
    ReceivedMessages = 97
    SentMessages = 207
    NumberOfReceivedSTMessages = 0
    NumberOfReceivedStatusMessages = 6
    NumberOfReceivedCommitMessages = 0
    LastExecutedSeqNumber = 46
    PrePrepareMessages = 37
    AvgBatchSize = 1.49
    AvgPendingRequest = 0.00
 [/concord/submodules/concord-bft/bftengine/src/bftengine/DebugStatistics.cpp:84]

That will help us find additional bottlenecks.

This also bumps the number of voters in the test to 10000. We were
getting through 1000 votes in less than ten seconds. That would mean
only one of the above stats would be printed during the test. This
will get us a few more readings over the test run.