Add an event recording utility for build/test scripts to use.
Usage:
python3 event_recorder.py record_event <stage_name> <event_name> <json_file>
python3 event_recorder.py summarize <json_file>

The record_event action will create a json file or update an existing json file.
The summarize action will create a pleasant looking summary.

The format of the json is:
{
  "stage1": {
    "events": {
      "event1": "2019-09-25_20-45-09",
      "event2": "2019-09-25_20-45-10",
      "event3": "2019-09-25_20-45-10",
      "event4": "2019-09-25_20-45-11"
    },
    "elapsed": "0:00:02"
  },
  "stage2": {
    "events": {
      "event1": "2019-09-25_20-45-10",
      "event2": "2019-09-25_20-45-10",
      "event3": "2019-09-25_20-45-11",
      "event4": "2019-09-25_20-45-11",
      "event5": "2019-09-25_20-45-11",
      "event6": "2019-09-25_20-45-12"
    },
    "elapsed": "0:00:02"
  },
  "stage3": {
    "events": {
      "event1": "2019-09-25_20-45-10",
      "event2": "2019-09-25_20-45-12",
      "event3": "2019-09-25_20-45-12"
    },
    "elapsed": "0:00:02"
  },
  "total": "0:00:03"
}

The "stage" is a big piece of the thing you are measuring.  The
"events" are smaller pieces.  The time for a stage is summarized,
as well as the time for the entire process being tracked.

A lock file for each json file is used to ensure multiple processes
can use this utility to record events about a process concurrently.