Replace busy wait and fight starvation in KVBClientPool
KVBClientPool has used a simple `while(true)` and `boost::yield` to
deal with more threads wanting concord-bft client-proxies than are
available, since it was created. This hasn't been a problem with
Ethereum clusters, because the number of threads
(api_client_worker_pool_size) has been less than the number of
client-proxies, so the yield was never called.

DAML, however, has many more API threads active than there are
client-proxies. When this yield loop gets invoked, it raises CPU usage
and reduces time available to other active threads.

While we may raise the number of client-proxies, we
probably still want to be able to have fewer of them than of API
threads.

This change replaces the yield-busy-wait with a condition
variable. This allows waiting threads to remain asleep until a client
is actually available.

At the same time, the yield-busy-wait could also lead to
starvation. Which thread got the next available client was arbitrary,
based on which one happened to try to pull from the queue next.

This change introduces a system to prevent starvation, by ensuring
that threads get clients in the order they started waiting.

Using Ethereum performance tests:

 1. Without this change, raising api_worker_pool_size above
    client_proxies_per_replica led to a decrease in throughput: 102tps
    during a time we were seeing ~130tps. Additionally, time spent
    during each phase of replica execution was higher on the primary
    than on other nodes.
    (https://blockchain.svc.eng.vmware.com/view/Blockchain%20on%20GitLab/job/Blockchain%20Performance%20Test%20on%20GitLab/207/)

 2. With this change, throughput stayed at least even, possibly even
    rose, to 140tps. Time in replica execution was more similar across
    nodes.
    (https://blockchain.svc.eng.vmware.com/view/Blockchain%20on%20GitLab/job/Blockchain%20Performance%20Test%20on%20GitLab/212/)

Some other side-effects:

* We can use std::queue instead of boost::lockfree::queue, since we
  hold a lock while modifying it. Not only is this simpler, but it
  will also remove one of our spurious valgrind warnings.

* The destructor of KVBClientPool now properly waits to destroy all
  clients before finishing. This may also reduce valgrind warnings.