Perform merge to TOT on build nodes before building/testing.
GitLab merge requests: Will always be merged.
Manual run with parameters: Will be merged if a checkbox is checked,
which it is by default.

This merge is not pushed.  It is just used for building and testing
a change.

This is a merge with --no-ff because that is what GitLab will do
when it merges the change for real.