VB-1727: Invitation retrieval not working for myVmware accounts
Root cause: the auth token csp issues has no reliable way to the get the email address of a user.  At the moment, this only matters in the invitation flow.

- Changed InvitationService to get the user info from CSP.  This includes the real email (the userName)
- Added a test for a bad user

TODO: This should be considered a workaround for a high priority bug.  We need to look into the whole handling of this in more detail.