Add Client Proxy ID-based Time Verification.
This commit adds a time sample verifcation scheme to the Concord time service based on client proxy IDs. This scheme is rooted in the idea that Concord-BFT should guarantee it is intractable to impersonate a client proxy without knowledge of a private key belonging to that client proxy, so we can validate time updates by making sure they originated with a client proxy on the same node as the time source they claim to originate from. This new time verification scheme is added as an alternative to the existing RSA signature-based time verification rather than a replacement. Using no time verification also remains an option. This commit introduces the time_verification top-level configuration parameter to select an alternative(which accepts a string value that can be either rsa-time-signing, bft-client-proxy-id, or none). To handle the implementation of this new alternative, TimeSigner and TimeVerifier have been generalized to abstract classes, with their existing implementations renamed RSATimeSigner and RSATimeVerifier. ClientProxyIDTimeVerifier is introduced as an alternative implementation of TimeVerifier (though it has no corresponding implementation of TimeSigner because the new ID-based time verification scheme does not directly use signatures). This commit also adds to the unit tests for the time service in order to test the new time verification method.