Retry builds when certain errors occur.
Store information about the build processes, look for errors due
to artifactory glitches, and if found, retry a couple times.  We
will look for 503 errors during a maven build and "unexpected EOF"
at the end of a log when pulling a Docker images.

This involves a rework of buildall.sh.

The code which monitors/evaluates/retries builds is isolated into
a separate library file.  The file buildall.sh defines the tasks,
launches them initially, and hands the processes over to the library.

Bash is a bit limited with regards to data structures, so we now
have four tables, each having the build name (like "concord-core")
as a key:
  BUILD_PROCESS_IDS # Build Name => Process ID
  BUILD_COMMANDS    # Build Name => Process command
  BUILD_FAILURES    # Build Name => Number of build failures
  BUILD_LOGS        # Build Name => Current log file for this build (changes with retries)

Since having four tables floating around is risky, I want as few
things as possible to write to them. These tables are only written
to by:
  retryFailures(): Writes to BUILD_FAILURES to increment the count.
  addToProcList(): Writes to all of them.

I've also broken things down into smaller functions for readability.