VB-1741: DBAdapter should own IDBClient
DBAdapter holds the IDBClient pointer passed to it in a shared_ptr, so
that when the DBAdapter is destructed, so is the IDBClient. This patch
remove's the unique_ptr that was wrapping the IDBClient in
run_service, so that we don't double-free the IDBClient.

This returns Concord to a non-segfaulting-on-exit state.