Cleanup Concord's Dockerfile's CMD for deployment.
This commit revises Concord's Dockerfile to change its CMD such that Concord's Docker image itself does not have a default command containing peculiarities of our docker-compose setup we use for development and testing. This should remove the need to give Concord a fake find-docker-instancs.sh file and/or an extra copy of its configuration in deployment to appease the commands we use with docker-compose that aren't needed in production. Everything other than the Concord launch itself is removed from the CMD in Concord's Dockerfile. Custom command values have been added for Concord containers in appropriate docker-compose*.yml files so that the extra commands we need to resolve container IPs when running with docker-compose will be run there. As part of this change, the memleak_util_cmd build argument to Concord's Dockerfile has also been removed in favor of adding the memleak utility command to a docker-compose*.yml file.