Pass signals to concord through docker correctly
This patch changes the concord CMD so that the concord process is run
directly, instead of being run under a bash shell. When run under a
shell, the SIGTERM sent by `docker stop` is swallowed by the shell
without concord seeing it. When run directly, concord catches the
signal and shuts down as expected.

This has a couple of nice properties:

1. Concord shuts down faster, because we don't have to wait for docker
to timeout and send SIGKILL.

2. We can test that concord exits with a non-error status code. This
helps us track down things like memory usage errors, because segfaults
cause non-zero exits.

A minimal regression test is added to enforce the simple case of
zero.

In the future, it might be nice to put the exit code assertion
directly in the general hermes shutdown steps. I haven't done that in
this change because in a quick test, concord2 segfaulted during
shutdown.