Make conc_genconfig statically linked.
This commit revises the build process for conc_genconfig in order to make it statically linked. The intention of this change is to make conc_genconfig a standalone Linux binary such that it can be placed in the Persephone Configuration Service image and run there without the need to install dependencies that don't come standard on (at least select common distribution(s) of) Linux. This commit also modifies persephone/config-service/Dockerfile to consume conc_genconfig as a static binary rather than building conc_genconfig from its sources and installing its dependencies.