DAML: Source envs from file
To make agent more agnostic to the application/container it is starting, this
change introduces a config file for the DAML Ledger API service. The config
file will be generated by config service and thereafter, forwarded until the
entrypoint.sh scripts reads it at start-up time. We ignore the file if it
doesn't exist and hence the normal dev/testing workflow won't be impacted.