Attempt simple auto-triage of build failures.
This covers builds, not tests.  When a build fails, the script knows the
associated file, and we know what typical errors look like.  So now the script
will carry out this process:
- While ignoring white space (could be an indented block),
- look for the word "error" (case insensitive),
- in the first 5 characters of the line (to include lines like "[Error] <message>"
  or leading with binary gibberish, while excluding things like names of source
  code files which contain the word "error"),
- and show that line plus the ten lines after it.

If a line matching the above is not found, the script will just print the last ten
lines of the file to the console.  Errors are usually at the end of a build log.

It will also tell the user which file to look at.