Modify agent to attach components to user-defined network
This patch promotes the use of user-defined container network to remove
the need to manually link containers together (--link is a legacy option
that will be removed in future versions of Docker).

The effect of using a non-default network is that container names are
automatically published via Docker DNS to facilitate simple (local)
service discovery.