ThinReplica: Client Scaffolding
This change introduces a new source directory for the thin replica client
library. The directory includes a sample library and application. Both are
integrated with log4cplus, the thin replica protobuf definition, gRPC and are
built with CMake. In addition, this change includes a Dockerfile to build and
run the sample application. With this change, most of the scaffolding code is
out of the way and we can focus on the acutal thin replica client
implementation.