Improve error reporting when Concord rejects its configuration.
This commit makes a number of changes intended to improve the errors and other feedback Concord generates in failure cases where it rejects its configuration file. In particular, these changes include: (1) The config/c command line option to Concord is made mandatory (this should cause boost::program_options to complain earlier and more explicitly if the argument is omitted). (2) A check is added to make initialize_config fail immediately and report an error if Concord's configuration file cannot be opened. (3) A check is added to make initialize_config fail immediately and report an error if Concord's configuration file appears to be empty. (4) Lists of missing parameters are added to exceptions that may trigger when Concord thinks needed parameters are missing; this is done to address an issue where, in certain environments/conditions (evidently including current Jenkins runs), log statements we were relying on to report the missing parameters failed to print and flush to output before a subsequent exception brings Concord down. (5) concord::config::detectLocalNode is modified to throw exceptions with messages describing common failure scenarios when applicable, as detectLocalNode's general-case error message can be somewhat cryptic if it shows up when a larger more obvious problem with the configuration exists. (6) Logic is added to have Concord's main function exit immediately if initialize_config detects that it failed to fully load Concord's configuration; this is intended to suppress any additional, possibly misleading or confusing, error messages or exceptions that may arise if it tries to continue running after initialize_config already reported a failure.