Remove unused cryptographic keys from Concord configuration.
It has been determined to be the case that the execution threshold cryptosystem and the RSA key pairs for Concord-BFT clients are not currently used by Concord-BFT, and that we have no plans to return to using them in the future. As such, this commit removes the disused cryptographic keys from Concord's configuration.