Give pytest suites a way to tell the framework to collect support bundles.
With this change, a test suite can write a file containing the info about
systems from which a support bundle needs to be collected.  When the suite
ends, the parent suite class will read the file and collect the appropriate
bundles.

This change also has CoreVMTests use this new functionality to collect
support bundles.