VB-2087: Add support to deploy multiple blockchains
This fixes on problem, and solves another.

Up to now, the backend enforced no limit on the number of blockchains that can be deployed, although the UI forced a limit of 1, unless you are the system admin.

The backend now enforces a limit.  A new org propterty has bee added called "maximum_chains"

The new rules are as follows
* System Operator can still deploy any number of blockchains
* maximum_chains == 0, organization can deploy any number (well, up to max integer)
* maximum_chains > 0, organization can deploy at most that many chains
* maximum_chains < 0, organization cannot deploy any blockchains
* maximum_chains missing, organization can deploy 1 blockchain

Changes as follows
* Constants: added definition of "maximum_chains"
* BlockchainController: logic to limit blockchains
* BlockchainControllerTest: Test the limits
* ErrorCodeType: added new error code for exceding the limit

Also a refactoring of the following files
* message.properties, message_en_us.properties: reorganized to alphatize entries and remove duplicate definitions
* ErrorCodeType: Alphabetized the declarations
* BadRequestException: Fixed a bug that was setting return status to 500

Fixing a bug.

Resolving VB-2120 along with this checkin.

merging changes to blockchaincontroller