Move Helen role tests to their own suite.
The role tests intermittently start experiencing a delay of multiple minutes
per call to CSP.  It may be due to rate limiting; we do not know.  So they
are being moved to their own suite.  After this checkin, a nightly run
will be set up so these tests are still run regularly, but in isolation.